LangServe - Language Translator/requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
pypdf
bs4
arxiv
fitz
frontend
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
streamlit
langchain_ollama
langchain_groq
langchain_core
fastapi
langserve
sse_starlette